python/requirements.txt
PyYAML==3.11
ckanapi==3.6
docopt==0.6.2
et-xmlfile==1.0.1
httplib2==0.9.2
jdcal==1.2
oauth2==1.9.0.post1
openpyxl==2.3.5
python-frontmatter==0.2.1
requests==2.10.0
simplejson==3.8.2
six==1.10.0
tablib==0.11.2
wsgiref==0.1.2
